% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplotwithlm.R
\name{dotplotwithlm}
\alias{dotplotwithlm}
\title{Dot Plot with LM abline}
\usage{
dotplotwithlm(inputdata, x, y)
}
\arguments{
\item{inputdata}{The data frame used for the dotplot and linear model.}

\item{x}{The x-axis variable}

\item{y}{The y-axis variable}
}
\value{
Outputs a dotplot with the linear model drawn onto it.
}
\description{
Dot Plot with LM abline
}
\examples{
\dontrun{spruce.df<-read.csv("Spruce.df");dotplotwithlm(inputdata=spruce.df, x=BHDiameter, y=Height)}
}
