% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial simulation over multiple iterations}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of simulations to run}

\item{n}{The number of elements}

\item{p}{The probability of success}
}
\value{
Outputs a barplot and a table with the data from the simulation.
}
\description{
Binomial simulation over multiple iterations
}
\examples{
\dontrun{mybin(100, 10, .7)}
}
