% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{A function that generates a 95 percent confidence interval for the population mean}
\usage{
myci(x)
}
\arguments{
\item{x}{The sample to generate the confidence interval from.}
}
\value{
The 95 percent confidence interval values.
}
\description{
A function that generates a 95 percent confidence interval for the population mean
}
\examples{
\dontrun{set.seed(23);x = rnorm(30,mean=10,sd=12);myci(x)}
}
