% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{A function to find the maximum likelihood estimate for a parameter.}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The function to apply to the generated matrix of all values.}

\item{x}{The data points.}

\item{param}{The parameter values to consider.}

\item{...}{Extra arguments to send to plot functions.}
}
\value{
A list of all related data, and a graph representing the maximum likelihood estimate.
}
\description{
A function to find the maximum likelihood estimate for a parameter.
}
\examples{
\dontrun{mymaxlik(x=c(3,3,4,3,4,5,5,4), param=seq(0,1, length=1000), lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
