% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal curve from negative infinity to a.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution.}

\item{sigma}{The standard deviation of the normal distribution.}

\item{a}{The x value that you want to draw the curve up to.}
}
\value{
Outputs a graph with the normal curve shaded from negative infinity to a and the area in that interval.
}
\description{
Normal curve from negative infinity to a.
}
\examples{
\dontrun{myncurve(0, 1, .5)}
}
