% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytstat.R
\name{mytstat}
\alias{mytstat}
\title{A function which outputs the t statistic for given sample data.}
\usage{
mytstat(z)
}
\arguments{
\item{z}{The sample data to return the t statistic for.}
}
\value{
The t-statistic for the sample data.
}
\description{
A function which outputs the t statistic for given sample data.
}
\examples{
\dontrun{mytstat(c(1,2,3,4,5))}
}
